analytics/pyproject.toml
[project]
name = "safe-analytics"
version = "0.1.0"
description = "Agent-native BI for the Safe indexer's ClickHouse sink"
requires-python = ">=3.10"
dependencies = [
    "clickhouse-connect>=0.8.0",
    "python-dotenv>=1.0.0",
    "pandas>=2.0.0",
    "matplotlib>=3.8.0",
    "tabulate>=0.9.0",
]
